lora/requirements_megaman_lora.txt
# Mega Man LoRA 訓練依賴項
# 核心深度學習框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion 模型相關
diffusers>=0.21.0
transformers>=4.25.0
accelerate>=0.20.0

# LoRA 和模型微調
peft>=0.5.0

# 圖像處理
Pillow>=9.0.0
opencv-python>=4.7.0

# 數據處理和工具
datasets>=2.0.0
numpy>=1.21.0
pandas>=1.3.0

# 進度條和日誌
tqdm>=4.64.0
tensorboard>=2.10.0

# 性能優化 (可選但推薦)
xformers>=0.0.16  # 需要 CUDA 支持
bitsandbytes>=0.35.0  # 8-bit 優化

# 其他工具
safetensors>=0.3.0
huggingface-hub>=0.15.0
wandb>=0.13.0  # 可選：實驗追蹤

# 開發工具
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0 